Fix ft_bzero.c, ft_memcpy.c, ft_memset.c, ft_striteri.c and Makefile

+++ common_core/libft/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: gsantama <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/29 15:21:19 by gsantama          #+#    #+#             */
-/*   Updated: 2024/11/29 15:21:21 by gsantama         ###   ########.fr       */
+/*   Updated: 2024/11/30 18:19:59 by gsantama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	if (!dst && !src)
 		return (0);
